Address Issue #13
cardboardlint is a proper package (__init__.py is not empty) because I
wanted to access the linters without doing some sort of path work
around. If there is a better way, feel free to change.

cardboardlinter stil has quite a bit of code because it contains
parse_args and run_diff