Fix bug in matches_filefilter
The function `matches_filefilter` would exit after encounter a rule that
matches the given file. However, if there are multiple rules that can be
satisfied simultaneously, for example both '+ *py' and '- x.py' would be
satisfied for file 'x.py', then `matches_filefilter` will provided
different result depending on the order of the rules.

Added change ensures that all of the rules are satisfied, rather than
the first. Tests and docstrings are updated.